Int -> Float
resolves https://github.com/orgs/codex-team/projects/10#card-47980015

+++ src/typeDefs/plans.ts
@@ -18,7 +18,7 @@ export default gql`
     """
     Monthly charge for plan
     """
-    monthlyCharge: Int!
+    monthlyCharge: Float!
 
     """
     Events limit for plan
